+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.2.0] - 2017-12-15
+### Added
+* Convenient conversions for `glutin::Context`, `winapi::D3D11Device`,
+  `winapi::D3D12Device`, and `winapi::windef::HGLRC` into
+  RenderDoc `DevicePointer`.
+
+### Changed
+* Update existing dependencies.
+* Optionally depend on `glutin` in place of `winit`.
+* Depend on `wio` for Windows targets.
+
+### Fixed
+* Missing byte in `SHADER_MAGIC_DEBUG_VALUE_STRUCT` broke Windows builds.
+
 ## 0.1.0 - 2017-10-11
 ### Added
 * Initial crate release.
@@ -13,4 +27,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Type-safe version requests and downgrading.
 * Convenient conversions for `winit::VirtualKeyCode` into RenderDoc `InputButton`.
 
-[Unreleased]: https://github.com/ebkalderon/renderdoc-rs/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/ebkalderon/renderdoc-rs/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/ebkalderon/renderdoc-rs/compare/v0.1.0...v0.2.0